# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Simple GoLang web server
  description: It is a simple web server written in the Go programming language.
    The project uses DDD and CQRS architecture.
  contact:
    name: Konstantin (Forzend)
    url: https://t.me/F0rzend
  license:
    name: MIT
    url: https://mit-license.org/
  version: 0.1.0
servers:
  - url: 'http://localhost:8080'
    description: Local debug server
...
paths:
  /users:
    post:
      tags:
        - Users
      summary: User sign up
      description: Create new user
      operationId: createUser
      requestBody:
        description: User registration data
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: F0rzend
                username:
                  type: string
                  example: forzend
                email:
                  type: string
                  format: email
                  example: some@gmail.com
        required: true

      responses:
        204:
          description: Location of created user
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: /users/1
        400:
          description: Validation error
        409:
          description: User already exists
        500:
          description: Error on server side
...
  /bitcoin:
    get:
      tags:
        - Finances
      summary: Get current bitcoin price
      description: Get current bitcoin price in USD
      responses:
        200:
          description: Current bitcoin price and last updating date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bitcoin'

    put:
      tags:
        - Finances
      summary: Change the bitcoin price
      description: Set new bitcoin price in USD
      requestBody:
        description: New bitcoin price
        content:
          application/json:
            schema:
              type: object
              properties:
                price:
                  type: number
                  format: float64
                  example: 100.00
      responses:
        204:
          description: Location to get updated bitcoin price
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: /bitcoin
        400:
          description: Validation error
components:
  schemas:
...
    Bitcoin:
      type: object
      properties:
        price:
          type: number
          format: float
          example: 100.00
        updated_at:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00Z
...